+++ utils.go
@@ -1,0 +1,7 @@
+package siwe
+
+import "strings"
+
+func isEmpty(str *string) bool {
+	return str == nil || len(strings.TrimSpace(*str)) == 0
+}
